+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 
+- Fix mobile devices duplication during mobile sync
 - Fix leftover data (profile rights and config values) after plugin uninstall
 
 ## [3.2.2] - 2026-08-03